Add and use `container-apiserver start`.
- Part of #653.
- Also use same file organization for network helper.
