+++ Sources/ContainerCommands/System/SystemStart.swift
@@ -64,6 +64,7 @@ extension Application {
                 args.append("--debug")
             }
 
+            args.append("start")
             let apiServerDataUrl = appRoot.appending(path: "apiserver")
             try! FileManager.default.createDirectory(at: apiServerDataUrl, withIntermediateDirectories: true)
             var env = ProcessInfo.processInfo.environment.filter { key, _ in
